chore(cli): remove the standalone `gortex index` command
`gortex index` built a throwaway in-memory graph, printed stats, and
exited — it never registered with the daemon, never wrote the shared
store, and never tracked anything. That made it useless for the
interactive flow it implies (a user runs `gortex index .`, then `query`
fails with "the daemon does not track …") and it diverged from how the
daemon actually indexes. The daemon-integrated path is `gortex track`.

Remove the command (index.go) and its one-shot summary helpers, keeping
the shared `humanizeInt` in a neutral humanize.go. The snapshot
machinery (saveSnapshotTo) stays — it is shared with the daemon's own
snapshot writer; only the `index --snapshot` entry point is dropped.

Also de-misinform: the inactive-daemon message no longer suggests
`gortex index .` as an activation path (it isn't one), and the
antigravity agent setup doc now points at `gortex track .`. Stale
`gortex index` mentions in comments, docs, and a retrieval fixture are
updated; the command group keeps `enrich`/`db`.
