+++ bench/fixtures/retrieval.yaml
@@ -400,7 +400,7 @@ cases:
       - cmd/gortex/mcp.go::mcpCmd
       - cmd/gortex/server.go::serverCmd
       - cmd/gortex/daemon.go::daemonCmd
-      - cmd/gortex/index.go::indexCmd
+      - cmd/gortex/track.go::trackCmd
       - cmd/gortex/eval.go::evalCmd
 
   - id: mh-config-load
